testing/web-platform/tests/webstorage/storage_session_setitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_session_setItem</h1>
    <div id="log"></div>
    <script>
        test(function() {
            sessionStorage.clear();
            sessionStorage.setItem("name", "user1");
            assert_equals(sessionStorage.length, 1, "localStorage.setItem")
        });
    </script>
 </body>
</html>
